+++ src/Ema/Generate.hs
@@ -6,9 +6,9 @@ module Ema.Generate where
 
 import Control.Exception (throw)
 import Control.Monad.Logger
-import Ema.Class
+import Ema.Class (Ema (staticAssets, staticRoutes), MonadEma)
 import Ema.Route (routeFile)
-import System.Directory (copyFile, createDirectoryIfMissing, doesDirectoryExist, doesFileExist)
+import System.Directory (copyFile, createDirectoryIfMissing, doesDirectoryExist, doesFileExist, doesPathExist)
 import System.FilePath (takeDirectory, (</>))
 import System.FilePattern.Directory (getDirectoryFiles)
 
@@ -24,7 +24,7 @@ generate ::
   m ()
 generate dest model render = do
   unlessM (liftIO $ doesDirectoryExist dest) $ do
-    error "Destination does not exist"
+    error $ "Destination does not exist: " <> toText dest
   let routes = staticRoutes model
   log LevelInfo $ "Writing " <> show (length routes) <> " routes"
   forM_ routes $ \r -> do
@@ -35,7 +35,11 @@ generate dest model render = do
       createDirectoryIfMissing True (takeDirectory fp)
       writeFileLBS fp s
   forM_ (staticAssets $ Proxy @route) $ \staticPath -> do
-    copyDirRecursively staticPath dest
+    liftIO (doesPathExist staticPath) >>= \case
+      True ->
+        copyDirRecursively staticPath dest
+      False ->
+        log LevelWarn $ toText $ "? " <> staticPath <> " (missing)"
 
 newtype StaticAssetMissing = StaticAssetMissing FilePath
   deriving (Show, Exception)